Add documentation files
This commit adds the following documentation files:
- SECURITY.md
- CHANGELOG.rst
- LICENSE
- CONTRIBUTING.md
- CODE_OF_CONDUCT.md

These files are included in the documentation using the `mdinclude` and `include` directives.

The commit also includes changes to the Sphinx configuration file `conf.py` to keep the auto-generated API documentation files and to exclude the `test_main` module from the documentation build.

Lastly, the commit adds the necessary CSS and JavaScript files for the Graphviz extension.